+++ apps/api/src/app/admin/admin.module.ts
@@ -11,6 +11,7 @@ import { Module } from '@nestjs/common';
 
 import { AdminController } from './admin.controller';
 import { AdminService } from './admin.service';
+import { QueueModule } from './queue/queue.module';
 
 @Module({
   imports: [
@@ -21,6 +22,7 @@ import { AdminService } from './admin.service';
     MarketDataModule,
     PrismaModule,
     PropertyModule,
+    QueueModule,
     SubscriptionModule,
     SymbolProfileModule
   ],