+++ libs/common/src/lib/interfaces/admin-jobs.interface.ts
@@ -1,0 +1,5 @@
+import { Job } from 'bull';
+
+export interface AdminJobs {
+  jobs: Job<any>[];
+}